+++ src/Middleware/AssetMiddleware.php
@@ -35,20 +35,20 @@ class AssetMiddleware implements MiddlewareInterface
      * @param \Psr\Http\Message\ServerRequestInterface $request The request
      * @param \Psr\Http\Server\RequestHandlerInterface $handler Request handler
      * @return \Psr\Http\Message\ResponseInterface A response
-     * @throws \Assets\Http\Exception\AssetNotFoundException
+     * @throws \Assets\Http\Exception\AssetNotFoundException|\Throwable
      */
     public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         /** @var \Cake\Http\ServerRequest $request */
         $file = Filesystem::instance()->concatenate(Configure::read('Assets.target'), $request->getParam('filename'));
-        Exceptionist::isReadable($file, __d('assets', 'File `{0}` doesn\'t exist', $file), AssetNotFoundException::class);
+        Exceptionist::isReadable($file, __d('assets', "File `{0}` doesn't exist", $file), AssetNotFoundException::class);
 
         $response = new Response();
         $response = $response->withModified(filemtime($file) ?: 0);
         /**
-         * @todo to be removed in a later version
+         * @todo to be removed in a later version, when CakePHP version is at least 4.4
          */
-        $method = version_compare(Configure::version(), '4.4', '>=') ? 'isNotModified' : 'checkNotModified';
+        $method = method_exists($request, 'isNotModified') ? 'isNotModified' : 'checkNotModified';
         if ($response->$method($request)) {
             return $response->withNotModified();
         }