+++ src/TestSuite/TestCase.php
@@ -35,12 +35,13 @@ abstract class TestCase extends CakeTestCase
     {
         parent::setUp();
 
-        @$this->loadPlugins(['Assets' => []]);
+        $this->loadPlugins(['Assets' => []]);
     }
 
     /**
      * Called after every test method
      * @return void
+     * @throws \Throwable
      */
     public function tearDown(): void
     {