+++ tests/TestCase/I18nTest.php
@@ -30,6 +30,6 @@ class I18nTest extends TestCase
     public function testI18nConstant(): void
     {
         $translator = I18n::getTranslator('assets', 'it');
-        $this->assertEquals('Il file `{0}` non esiste', $translator->translate('File `{0}` doesn\'t exist'));
+        $this->assertEquals('Il file `{0}` non esiste', $translator->translate("File `{0}` doesn't exist"));
     }
 }