+++ tests/test_app/TestApp/templates/Error/error500.php
@@ -1,10 +1,14 @@
+<?php
+use Cake\Core\Configure;
+/** @var string $message */
+?>
 <h2><?= __d('cake', 'An Internal Error Has Occurred.') ?></h2>
 <p class="error">
     <strong><?= __d('cake', 'Error') ?>: </strong>
     <?= h($message) ?>
 </p>
 <?php
-if (\Cake\Core\Configure::read('debug')) :
+if (Configure::read('debug')) :
     echo $this->element('auto_table_warning');
     echo $this->element('exception_stack_trace');
 endif;